refactor(parse-cargo): extract AI helpers to lib/parsing/parse-cargo-ai
buildCargoPrompts + parseCargoAIResponse + RawCargoItem moved out of the
route file so an offline script can import them without dragging in the
Next.js runtime. Behavior-preserving; route re-exports parseCargoAIResponse
for existing test importers. Mirrors the parse-vessel-helpers layout.
